# api/openapi.yaml
openapi: 3.0.3
info:
  title: ForgeC2 Command & Control API
  description: |
    ForgeC2 is a professional-grade command and control framework for red team operations.

    ## Authentication
    Most endpoints require a session cookie (`forgec2_session`) from `POST /login`.
    Unauthenticated access to `/api/*` returns **JSON 401** (`{success:false,error:...}`), not HTML.

    ## CSRF
    State-changing requests need double-submit CSRF: cookie `forgec2_csrf` + header `X-CSRF-Token`.

    ## Spec quality (honest)
    - Route inventory is kept near-complete (CI coverage ≥90%, stale=0).
    - **Core operator paths** (health, agents, modules, listeners, dashboard, deploy)
      include request/response schemas; many secondary routes remain summary-only.
    - Dual-use paths (e.g. `GET /agents`) serve SPA HTML or JSON depending on Accept/auth.
    - `/debug/pprof/*` is intentionally omitted (disabled by default).

    ## Rate Limiting
    - Login: Limited attempts with lockout
    - API: Token bucket rate limiting per user

    ## Permissions
    Role-based access control (custom roles supported):
    - **admin**: Full control (all permissions)
    - **user**: Standard operator permissions (agents, tasks, credentials, limited settings)

  version: 2.4.1
  contact:
    name: ForgeC2 Support
    email: support@forgec2.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
...
  - url: https://c2.example.com
...
paths:
...
  /agents/{id}/files/read:
    post:
      tags: [Files, Agents]
      summary: Read file
      description: Read a file's contents from the agent
      operationId: readFile
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: File path to read
      responses:
        '200':
          description: Read task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: forgec2_session
      description: Session cookie obtained from login endpoint
  schemas:
...
    TaskResponse:
      type: object
      required: [success, task_id]
      properties:
        success:
          type: boolean
          example: true
        task_id:
          type: integer
...